+++ cli/scripts/cluster.py
@@ -9,22 +9,11 @@ import pgbench, northwind
 base_dir = "cluster"
 
 
-def add_repset_tables(p_spk, p_repset, p_tbls, p_db, p_host, p_usr, p_key):
-  for tbl in p_tbls:
-    util.echo_cmd(p_spk + " repset-add-table " + p_repset + " " + tbl + " --db " + p_db,
-                  host=p_host, usr=p_usr, key=p_key)
-
-
-def drop_tables(p_tbls, p_nc, p_db, p_pg, p_host, p_usr, p_key):
-  for tbl in reversed(p_tbls):
-    util.psql_cmd("DROP TABLE " + tbl, p_nc, p_db, p_pg, host=p_host, usr=p_usr, key=p_key)
-
-
 def log_old_vals(p_run_sums, p_nc, p_db, p_pg, p_host, p_usr, p_key):
   for tbl_col in p_run_sums:
     tbl_col_lst = tbl_col.split(".")
-    tbl = tbl_col_lst[0]
-    col = tbl_col_lst[1]
+    tbl = tbl_col_lst[0] + "." + tbl_col_lst[1]
+    col = tbl_col_lst[2]
 
     cmd = "ALTER TABLE " + tbl + " ALTER COLUMN " + col + " SET (LOG_OLD_VALUE=true)"
     util.psql_cmd(cmd, p_nc, p_db, p_pg, p_host, p_usr, p_key)